"Librarize" common git cloning code across scripts
Both the merge forward script and build-all have had duplicated
code for a while. The introduction of docs-builder came with a
third place where git cloning was done. I put in a hacky fix in
centralizing the git cloning and update code.

This is that follow-up. :)

The diff is, I think, a case of "It looks worse than it is." The
basic plan I followed was to functionalize behaviors, and then move
any function that isn't "parse_args", "main", etc. up into the lib
modules to be shared by the multiple scripts that want to use them.

To test, build-all, docs-builder, and the update-version-and-merge-forward
scripts should all work as expected, and definitely not explode.
