+++ ci/build.py
@@ -28,7 +28,6 @@ parser.add_argument("--release", action="store_true", default=False,
 
 opts = parser.parse_args()
 
-
 builder.ensure_dir(WORKING_DIR, clean=True)
 TITO_DIR = os.path.join(WORKING_DIR, 'tito')
 MASH_DIR = os.path.join(WORKING_DIR, 'mash')
@@ -38,40 +37,27 @@ builder.ensure_dir(MASH_DIR, clean=True)
 # Initialize our connection to koji
 builder.init_koji()
 
-# Build our working_dir
-working_dir = WORKING_DIR
-print working_dir
+# Build our WORKING_DIR
+builder.ensure_dir(WORKING_DIR, clean=True)
 
-# Get the project to build from git
+# Get the project to build from git by spoofing a release config component
+component = {
+    'name': opts.project,
+    'git_url': "git@github.com:pulp/{}.git".format(opts.project),
+    'git_branch': opts.branch,
+}
+project_path = builder.clone_branch(component)
 
-git_repo = "git@github.com:pulp/{project}.git".format(project=opts.project)
-print "Getting git repo: {REPO}".format(REPO=git_repo)
-command = ['git', 'clone', git_repo, '--branch', opts.branch]
-subprocess.call(command, cwd=working_dir)
-
-print "Building list of things to build"
+print("Building list of things to build")
 
 download_list = []
 build_list = []
 
-# Check for external deps
-# for component in get_components(configuration):
-#     external_deps_file = component.get('external_deps')
-#     if external_deps_file:
-#         external_deps_file = os.path.join(working_dir, component.get('name'), external_deps_file)
-#         for package_nevra in builder.get_build_names_from_external_deps_file(external_deps_file):
-#             info = builder.mysession.getBuild(package_nevra)
-#             if info:
-#                 download_list.extend(builder.get_urls_for_build(builder.mysession, package_nevra, rpmsig=rpm_signature))
-#             else:
-#                 print "External deps requires %s but it could not be found in koji" % package_nevra
-#                 sys.exit(1)
-
 # Get all spec files
-for spec in builder.find_all_spec_files(os.path.join(working_dir, opts.project)):
+for spec in builder.find_all_spec_files(project_path):
     spec_nvr = builder.get_package_nvr_from_spec(spec)
     package_dists = builder.get_dists_for_spec(spec)
-    print "%s %s" % (spec_nvr, package_dists)
+    print("%s %s" % (spec_nvr, package_dists))
     for package_nevra in builder.get_package_nevra(spec_nvr, package_dists):
         info = builder.mysession.getBuild(package_nevra)
         if info:
@@ -85,17 +71,8 @@ download_list = sorted(download_list, key=lambda download: download[1])
 # Perform a scratch build of all the unbuilt packages
 build_ids = []
 
-# print download_list
-print build_list
-
 if build_list:
-    # if rpm_signature:
-    #     print "ERROR: rpm signature specificed but the following releases have not been built."
-    #     for spec, dist in build_list:
-    #         print "%s %s" % (spec, dist)
-    #     sys.exit(1)
-
-    print "Performing koji scratch build "
+    print("Performing koji scratch build")
     for spec, dist in build_list:
         spec_dir = os.path.dirname(spec)
         builder.build_srpm_from_spec(spec_dir, TITO_DIR, testing=True, dist=dist)
@@ -105,7 +82,7 @@ if build_list:
     builder.wait_for_completion(build_ids)
 
     if opts.release:
-        print "Performing koji release build"
+        print("Performing koji release build")
         # Clean out the tito dir first
         builder.ensure_dir(TITO_DIR)
         spec_dir_set = set()
@@ -114,10 +91,10 @@ if build_list:
             if spec_dir not in spec_dir_set:
                 spec_dir_set.add(spec_dir)
                 # make sure we are clean to merge forward before tagging
-                print "validating merge forward for %s" % spec_dir
                 git_branch = promote.get_current_git_upstream_branch(spec_dir)
                 parent_branch = component.get('parent_branch', None)
-                promotion_chain = promote.get_promotion_chain(spec_dir, git_branch, parent_branch=parent_branch)
+                promotion_chain = promote.get_promotion_chain(
+                    spec_dir, git_branch, parent_branch=parent_branch)
                 promote.check_merge_forward(spec_dir, promotion_chain)
                 # Tito tag the new releases
                 command = ['tito', 'tag', '--keep-version', '--no-auto-changelog']
@@ -137,7 +114,7 @@ if build_list:
             # Merge merge the commit forward, pushing along the way
             promote.merge_forward(spec_dir, push=True)
 
-print "Downloading rpms"
+print("Downloading rpms")
 # Download all the files
 builder.download_builds(MASH_DIR, download_list)
 builder.download_rpms_from_scratch_tasks(MASH_DIR, build_ids)
